+++ CHANGELOG.md
@@ -23,6 +23,9 @@ The version is represented by three digits: a.b.c.
 FEATURE:
 - `symmetria.Permutation`: add `ascents` method
 - `symmetria.CyclePermutation`: add `ascents` method
+- `symmetria.Permutation`: add `descents` method
+- `symmetria.CyclePermutation`: add `descents` method
+
 
 ## \[0.0.4\] - 2024-05-28